requirements.txt
Flask
Flask-CORS
git+https://github.com/openai/whisper.git
pinecone-client
pydub
sentence-transformers
langchain
torch
torchaudio
ffmpeg-python
